move services from client to shard to avoid circular dependency
